refactor: split synthetic run builders
